Add support for ical events
This also switches up the configuration of scrapers so that it's not
necessary to keep writing boilerplate scrapers as their own file. The
ical format is pretty simple but Google will normalize to UTC, which
goes against what normally is the case. Pulling in moment to deal with
the tz conversion... resisted long enough.